+++ lib/Data/Time/LocalTime/Internal/LocalTime.hs
@@ -7,6 +7,8 @@ module Data.Time.LocalTime.Internal.LocalTime
     -- * Local Time
     LocalTime(..),
 
+    addLocalTime,diffLocalTime,
+
     -- converting UTC and UT1 times to LocalTime
     utcToLocalTime,localTimeToUTC,ut1ToLocalTime,localTimeToUT1,
 ) where
@@ -20,7 +22,9 @@ import Data.Data
 #endif
 import Data.Time.Calendar.Days
 import Data.Time.Calendar.Gregorian
+import Data.Time.Clock.Internal.NominalDiffTime
 import Data.Time.Clock.Internal.UniversalTime
+import Data.Time.Clock.Internal.UTCDiff
 import Data.Time.Clock.Internal.UTCTime
 import Data.Time.LocalTime.Internal.TimeOfDay
 import Data.Time.LocalTime.Internal.TimeZone
@@ -48,6 +52,14 @@ instance NFData LocalTime where
 
 instance Show LocalTime where
     show (LocalTime d t) = (showGregorian d) ++ " " ++ (show t)
+
+-- | addLocalTime a b = a + b
+addLocalTime :: NominalDiffTime -> LocalTime -> LocalTime
+addLocalTime x = utcToLocalTime utc . addUTCTime x . localTimeToUTC utc
+
+-- | diffLocalTime a b = a - b
+diffLocalTime :: LocalTime -> LocalTime -> NominalDiffTime
+diffLocalTime a b = diffUTCTime (localTimeToUTC utc a) (localTimeToUTC utc b)
 
 -- | Get the local time of a UTC time in a time zone.
 utcToLocalTime :: TimeZone -> UTCTime -> LocalTime